@@ -6,10 +6,11 @@
 	"keywords": ["lcs"],
 	"license": "MIT",
 	"require": {
-		"php": ">=8.0"
+		"php": ">=8.1"
 	},
 	"require-dev": {
-		"phpunit/phpunit": "*"
+		"phpunit/phpunit": "^10.0",
+		"s9e/repdoc": "dev-master"
 	},
 	"autoload": {
 		"psr-4": {
@@ -20,5 +21,9 @@
 		"psr-4": {
 			"s9e\\LongestCommonSubstring\\Tests\\": "tests"
 		}
+	},
+	"scripts": {
+		"patch-docs": "vendor/bin/repdoc README.md",
+		"test": "vendor/bin/phpunit"
 	}
 }
